Code behind za New Usera prebačen sa 'LogInScreen.aspx.cs' na 'NewUser.aspx.cs'
