+++ PraProjekt/LogInScreen.aspx.cs
@@ -14,17 +14,6 @@ namespace PraProjekt
         {
         }
 
-        protected void btnRegister_Click(object sender, EventArgs e)
-        {
-            Repository.CreateUserAcc(new UserAcc()
-            {
-                Email = tbEmail.Text,
-                Pass = tbPass.Text,
-                Username = tbUserName.Text,
-                IsActive = 1
-            });
-            
-            Response.Redirect("CreateQuiz.aspx");
-        }
+        
     }
 }
